J2N.Numerics.BitOperation::TrailingZeroCount(this int value): Added missing casts to Unsafe TrailingZeroCount optimization to make the value unsigned.

+++ src/J2N/Numerics/BitOperation.cs
@@ -399,7 +399,7 @@ namespace J2N.Numerics
                 // Using deBruijn sequence, k=2, n=5 (2^5=32) : 0b_0000_0111_0111_1100_1011_0101_0011_0001u
                 ref MemoryMarshal.GetReference(TrailingZeroCountDeBruijn),
                 // uint|long -> IntPtr cast on 32-bit platforms does expensive overflow checks not needed here
-                (IntPtr)(int)(((value & (uint)-(int)value) * 0x077CB531u) >> 27)); // Multi-cast mitigates redundant conv.u8
+                (IntPtr)(int)(((((uint)value) & (uint)-(int)(uint)value) * 0x077CB531u) >> 27)); // Multi-cast mitigates redundant conv.u8
 #endif
         }
